SiteVisit-noatts
Needs refreshing or deleting. Summary of activities that may be associated with a site-visit. This is an old diagram from ANSNIS info model 1.5. It was removed from the word document where it was replaced with the up to date diagram https://github.com/ANZSoilData/def-au-domain/blob/main/fig/SiteVisit.png